feat: reject circles that are shrinking segments
When driving through an aisle or entering an aisle, we will first see segment type obstacles. As the wall starts going out of view, these are then (incorrectly) classified as circle type obstacles, which end up shrinking, causing their centers to move toward the robot, causing false hard stops. This feature detects when a circle appears close to where a segment previously was in a manner where the endpoints are close on one end and the other end is colinear with the old segment. If the criteria are met, the conversion to a cicle is rejected.